requirements.txt
# prod
Django==5.1
djangorestframework==3.15
psycopg2-binary==2.9.9
Pillow==10.4.0
django-cors-headers==4.4.0
django-environ==0.11.2
django_simple_captcha==0.6.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
bleach==6.1.0
channels[daphne]

# dev
black==24.8.0
pre_commit==3.8.0
django-silk==5.2.0